+++ R/add.R
@@ -1069,8 +1069,8 @@ e_sankey.echarts4rProxy <- function(e, source, target, value, layout = "none", r
 #' @param name Name of graph.
 #' @param nodes Data.frame of nodes.
 #' @param names Names of nodes, unique.
-#' @param value values of nodes.
-#' @param size Size of nodes.
+#' @param value Values of nodes or edges.
+#' @param size Sizes of nodes or edges.
 #' @param symbol Symbols of nodes.
 #' @param legend Whether to add serie to legend.
 #' @param category Group of nodes (i.e.: group membership).
@@ -1094,16 +1094,20 @@ e_sankey.echarts4rProxy <- function(e, source, target, value, layout = "none", r
 #'   stringsAsFactors = FALSE
 #' )
 #'
+#' value_edges <- sample(1:100, 20, replace = TRUE)
 #' edges <- data.frame(
 #'   source = sample(nodes$name, 20, replace = TRUE),
 #'   target = sample(nodes$name, 20, replace = TRUE),
+#'   value = value_edges,
+#'   size = ceiling(value_edges/20),
 #'   stringsAsFactors = FALSE
 #' )
 #'
 #' e_charts() |>
 #'   e_graph() |>
 #'   e_graph_nodes(nodes, name, value, size, grp, symbol) |>
-#'   e_graph_edges(edges, source, target)
+#'   e_graph_edges(edges, source, target, value, size) |>
+#'   e_tooltip()
 #'
 #' # Use graphGL for larger networks
 #' nodes <- data.frame(
@@ -1324,22 +1328,34 @@ e_graph_nodes.echarts4rProxy <- function(e, nodes, names, value, size, category,
 
 #' @rdname graph
 #' @export
-e_graph_edges <- function(e, edges, source, target) UseMethod("e_graph_edges")
+e_graph_edges <- function(e, edges, source, target, value, size) UseMethod("e_graph_edges")
 
 #' @method e_graph_edges echarts4r
 #' @export
-e_graph_edges.echarts4r <- function(e, edges, source, target) {
+e_graph_edges.echarts4r <- function(e, edges, source, target, value, size) {
   if (missing(edges) || missing(source) || missing(target)) {
     stop("must pass edges, source and target", call. = FALSE)
   }
-
+  
+  if (missing(size)) {
+    size <- NULL
+  }
+  
+  if (missing(value)) {
+    value <- NULL
+  }
+  
   source <- dplyr::enquo(source)
   target <- dplyr::enquo(target)
-
+  value <- dplyr::enquo(value)
+  size <- dplyr::enquo(size)
+  
   data <- .build_graph_edges(
     edges,
     source,
-    target
+    target,
+    value,
+    size
   )
 
   # build JSON data
@@ -1350,18 +1366,26 @@ e_graph_edges.echarts4r <- function(e, edges, source, target) {
 
 #' @method e_graph_edges echarts4rProxy
 #' @export
-e_graph_edges.echarts4rProxy <- function(e, edges, source, target) {
+e_graph_edges.echarts4rProxy <- function(e, edges, source, target, value, size) {
   if (missing(edges) || missing(source) || missing(target)) {
     stop("must pass edges, source and target", call. = FALSE)
   }
-
+  
+  if (missing(size)) {
+    size <- NULL
+  }
+  
   source <- dplyr::enquo(source)
   target <- dplyr::enquo(target)
-
+  value <- dplyr::enquo(value)
+  size <- dplyr::enquo(size)
+  
   data <- .build_graph_edges(
     edges,
     source,
-    target
+    target,
+    value,
+    size
   )
 
   # build JSON data